BAH-4430: Vivek: refactor(Allergies): extract no-known-allergy fetch into custom hook
